refactor(tui): move gitdiff under presentation/tui
gitdiff is a diff-parsing data layer for the /diff panel - its own package doc
calls it "a small, focused data layer behind the /diff changes panel". Every
consumer of its ~700 lines of parsing (Source, ReadSource, FileChange, Hunk,
FilePatch, Status, SplitFilePatchHunk, IsRepo) is under presentation/tui. It
was never platform-shared; it only looked shared because one 11-line git-exec
helper sat in the same file.

Split that helper out first: RunGit moves to platform/utils, next to
ProcessAlive, as the shared shell-out. internal/agent/agent_utils.go used
gitdiff for RunGit and nothing else, so its gitdiff import disappears entirely -
which is what allows the rest to move under presentation, where the agent is
forbidden by depguard from importing it.

gitdiff itself now calls utils.RunGit, so there is one definition, not a copy.
TestRunGit moves with the function.
