+++ internal/platform/memory/git.go
@@ -13,11 +13,11 @@ import (
 )
 
 // GitBackend syncs the memory directory with a git remote. It shells out to the
-// git CLI (mirroring the exec.Command("git", ...) style in internal/platform/
-// gitdiff) and inherits the ambient environment unchanged, so auth uses the
-// user's default git/ssh config (ssh-agent, credential helper, GIT_* env). Every
-// command runs under a per-operation timeout so a misconfigured remote ends the
-// command instead of hanging on a credential prompt.
+// git CLI (mirroring the exec.Command("git", ...) style in platform/utils.RunGit)
+// and inherits the ambient environment unchanged, so auth uses the user's default
+// git/ssh config (ssh-agent, credential helper, GIT_* env). Every command runs
+// under a per-operation timeout so a misconfigured remote ends the command
+// instead of hanging on a credential prompt.
 //
 // All operations are best-effort: they return an error for tests/telemetry, but
 // callers log and continue - a sync failure never aborts the agent run.
