+++ internal/presentation/tui/app/chat.go
@@ -25,7 +25,6 @@ import (
 	convdomain "github.com/inference-gateway/cli/internal/conversation/domain"
 	constants "github.com/inference-gateway/cli/internal/platform/constants"
 	formatting "github.com/inference-gateway/cli/internal/platform/formatting"
-	gitdiff "github.com/inference-gateway/cli/internal/platform/gitdiff"
 	logger "github.com/inference-gateway/cli/internal/platform/logger"
 	storage "github.com/inference-gateway/cli/internal/platform/storage"
 	shortcuts "github.com/inference-gateway/cli/internal/presentation/shortcuts"
@@ -33,6 +32,7 @@ import (
 	autocomplete "github.com/inference-gateway/cli/internal/presentation/tui/autocomplete"
 	components "github.com/inference-gateway/cli/internal/presentation/tui/components"
 	factory "github.com/inference-gateway/cli/internal/presentation/tui/components/factory"
+	gitdiff "github.com/inference-gateway/cli/internal/presentation/tui/gitdiff"
 	handlers "github.com/inference-gateway/cli/internal/presentation/tui/handlers"
 	keybinding "github.com/inference-gateway/cli/internal/presentation/tui/keybinding"
 	styles "github.com/inference-gateway/cli/internal/presentation/tui/styles"
