+++ internal/presentation/tui/handlers/chat_shortcut_handler.go
@@ -14,9 +14,9 @@ import (
 	agentdomain "github.com/inference-gateway/cli/internal/agent/domain"
 	conversation "github.com/inference-gateway/cli/internal/conversation"
 	convdomain "github.com/inference-gateway/cli/internal/conversation/domain"
-	gitdiff "github.com/inference-gateway/cli/internal/platform/gitdiff"
 	logger "github.com/inference-gateway/cli/internal/platform/logger"
 	shortcuts "github.com/inference-gateway/cli/internal/presentation/shortcuts"
+	gitdiff "github.com/inference-gateway/cli/internal/presentation/tui/gitdiff"
 	icons "github.com/inference-gateway/cli/internal/presentation/tui/styles/icons"
 )
